+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/Analytics/AnalyticsEventHandler.cs
@@ -17,7 +17,7 @@ namespace Stormancer.Server.Plugins.Users.Analytics
         }
         public Task OnLoggedIn(LoginContext loginCtx)
         {
-            analytics.Push("user", "login", JObject.FromObject(new { SessionId = loginCtx.Session.SessionId.ToString(), UserId = loginCtx.Session.User?.Id, PlatformId = loginCtx.Session.platformId }));
+            analytics.Push("user", "login", JObject.FromObject(new { SessionId = loginCtx.Session.SessionId.ToString(), UserId = loginCtx.Session.User?.Id, PlatformId = loginCtx.Session.platformId,  loginCtx.Dimensions }));
             return Task.CompletedTask;
         }